@@ -1,4 +1,7 @@
-import sys
-from os.path import dirname
-sys.path.append(dirname(__file__))
-from termgraph import *
+__all__ = ["main"]
+
+def __getattr__(name):
+    if name == "main":
+        from .termgraph import main
+        return main
+    raise AttributeError(f"module '{__name__}' has no attribute '{name}'")
